+++ src/Component/JsonSchema/Guesser/Guess/NonObjectGuessInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Jane\Component\JsonSchema\Guesser\Guess;
+
+interface NonObjectGuessInterface
+{
+}