+++ src/Component/OpenApi31/Tests/fixtures/issue-963/expected/Model/MemberDetails.php
@@ -1,0 +1,169 @@
+<?php
+
+namespace Jane\Component\OpenApi31\Tests\Expected\Model;
+
+class MemberDetails
+{
+    /**
+     * @var array
+     */
+    protected $initialized = [];
+    public function isInitialized($property): bool
+    {
+        return array_key_exists($property, $this->initialized);
+    }
+    /**
+     * @var int
+     */
+    protected $id;
+    /**
+     * @var string
+     */
+    protected $firstname;
+    /**
+     * @var string
+     */
+    protected $lastname;
+    /**
+     * @var string
+     */
+    protected $surname;
+    /**
+     * @var string
+     */
+    protected $description;
+    /**
+     * @var string
+     */
+    protected $email;
+    /**
+     * @var string
+     */
+    protected $role;
+    /**
+     * @return int
+     */
+    public function getId(): int
+    {
+        return $this->id;
+    }
+    /**
+     * @param int $id
+     *
+     * @return self
+     */
+    public function setId(int $id): self
+    {
+        $this->initialized['id'] = true;
+        $this->id = $id;
+        return $this;
+    }
+    /**
+     * @return string
+     */
+    public function getFirstname(): string
+    {
+        return $this->firstname;
+    }
+    /**
+     * @param string $firstname
+     *
+     * @return self
+     */
+    public function setFirstname(string $firstname): self
+    {
+        $this->initialized['firstname'] = true;
+        $this->firstname = $firstname;
+        return $this;
+    }
+    /**
+     * @return string
+     */
+    public function getLastname(): string
+    {
+        return $this->lastname;
+    }
+    /**
+     * @param string $lastname
+     *
+     * @return self
+     */
+    public function setLastname(string $lastname): self
+    {
+        $this->initialized['lastname'] = true;
+        $this->lastname = $lastname;
+        return $this;
+    }
+    /**
+     * @return string
+     */
+    public function getSurname(): string
+    {
+        return $this->surname;
+    }
+    /**
+     * @param string $surname
+     *
+     * @return self
+     */
+    public function setSurname(string $surname): self
+    {
+        $this->initialized['surname'] = true;
+        $this->surname = $surname;
+        return $this;
+    }
+    /**
+     * @return string
+     */
+    public function getDescription(): string
+    {
+        return $this->description;
+    }
+    /**
+     * @param string $description
+     *
+     * @return self
+     */
+    public function setDescription(string $description): self
+    {
+        $this->initialized['description'] = true;
+        $this->description = $description;
+        return $this;
+    }
+    /**
+     * @return string
+     */
+    public function getEmail(): string
+    {
+        return $this->email;
+    }
+    /**
+     * @param string $email
+     *
+     * @return self
+     */
+    public function setEmail(string $email): self
+    {
+        $this->initialized['email'] = true;
+        $this->email = $email;
+        return $this;
+    }
+    /**
+     * @return string
+     */
+    public function getRole(): string
+    {
+        return $this->role;
+    }
+    /**
+     * @param string $role
+     *
+     * @return self
+     */
+    public function setRole(string $role): self
+    {
+        $this->initialized['role'] = true;
+        $this->role = $role;
+        return $this;
+    }
+}